Update Intern.java
The UML diagram says for the Intern class to extend StaffMember, however the code says to extend StaffEmployee... in order to not have to hardcode the payrate when instantiating an intern, the intern class should extend staff member since it has no pay. 

Thinking about this in real life, interns are staff members, they are not employees.